FP-324, FP-338 Добавлена новая фича: при попытке ухода со страницы загрузки файлов, если есть незаконченные загрузки, вывется окно для подтверждения с результатами загрузки на текущий момент. Сделаны косметические доработки для справочника файлов правил расчета

+++ trunk/tc/src/org/flexpay/tc/actions/tariff/TariffCalcRulesFileDeleteAction.java
@@ -15,6 +15,11 @@ public class TariffCalcRulesFileDeleteAction extends FPActionWithPagerSupport<Ta
 	@NotNull
 	public String doExecute() {
 
+		if (id == null || id <= 0) {
+			addActionError(getText("tc.error.incorrect_rule_id"));
+			return REDIRECT_SUCCESS;
+		}
+
 		tariffCalculationRulesFileService.delete(id);
 
 		return REDIRECT_SUCCESS;
